# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Studio Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Current user auth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"
  /login:
    get:
      tags: [auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint.
...
  /logout:
    get:
      tags: [auth]
      operationId: logoutBrowserSession
      summary: Clear the session
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "302":
          description: Redirect to OIDC logout.
  /mobile-auth/token-exchange:
    post:
      tags: [auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileTokenExchangeRequest"
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /mobile-auth/logout:
    post:
      tags: [auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        "200":
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutMobileSessionResponse"
...
  /allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List all allocations
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
        - name: weekStart
          in: query
          schema:
            type: string
            format: date
        - name: weekEnd
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allocation"
    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Create a resource allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationRequest"
      responses:
        "201":
          description: Created allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
  /allocations/{id}:
    put:
      operationId: updateAllocation
      tags: [allocations]
      summary: Update an allocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllocationRequest"
      responses:
        "200":
          description: Updated allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
    delete:
      operationId: deleteAllocation
      tags: [allocations]
      summary: Delete an allocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Allocation deleted
...
  /utilization:
    get:
      operationId: getUtilization
      tags: [allocations]
      summary: Get team utilization data
      parameters:
        - name: weekStart
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: weekEnd
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Team utilization summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserUtilization"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    GetCurrentAuthUserResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required:
        - authenticated
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
...
    ExchangeMobileAuthorizationCodeResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutMobileSessionResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        replitId:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
          enum: [designer, pm, admin]
        hourlyRate:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - replitId
        - username
        - role
        - createdAt
...
    Allocation:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        projectColor:
          type: string
          nullable: true
        phaseId:
          type: string
          nullable: true
        phaseName:
          type: string
          nullable: true
        allocatedHours:
          type: number
        loggedHours:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - projectId
        - projectName
        - allocatedHours
        - loggedHours
        - startDate
        - endDate
        - createdAt
    CreateAllocationRequest:
      type: object
      properties:
        userId:
          type: string
        projectId:
          type: string
        phaseId:
          type: string
          nullable: true
        allocatedHours:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      required:
        - userId
        - projectId
        - allocatedHours
        - startDate
        - endDate
    UpdateAllocationRequest:
      type: object
      properties:
        allocatedHours:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
...
    UserUtilization:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        profileImage:
          type: string
          nullable: true
        allocatedHours:
          type: number
        loggedHours:
          type: number
        targetHours:
          type: number
        utilizationPercent:
          type: number
        status:
          type: string
          enum: [under, on_track, over]
        projects:
          type: array
          items:
            type: object
            properties:
              projectId:
                type: string
              projectName:
                type: string
              projectColor:
                type: string
                nullable: true
              allocatedHours:
                type: number
            required:
              - projectId
              - projectName
              - allocatedHours
      required:
        - userId
        - userName
        - allocatedHours
        - loggedHours
        - targetHours
        - utilizationPercent
        - status
        - projects
...