# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  replitId: string;
  username: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  profileImage?: string | null;
  role: UserRole;
  hourlyRate?: number | null;
  createdAt: Date;
}